// src/modules/transfers/domain/transfer-status.enum.ts
export enum TransferStatus {
  INITIATED = 'INITIATED',
  DEBIT_IN_PROGRESS = 'DEBIT_IN_PROGRESS',
  DEBITED = 'DEBITED',
  CREDIT_IN_PROGRESS = 'CREDIT_IN_PROGRESS',
  COMPLETED = 'COMPLETED',
  FAILED = 'FAILED',
  COMPENSATION_IN_PROGRESS = 'COMPENSATION_IN_PROGRESS',
  COMPENSATED = 'COMPENSATED',
}